# DOCS/contracts/openapi.json
{
...
  "openapi": "3.0.1",
  "info": {
    "title": "ConfluxPay API",
    "description": "ConfluxPay backend API â€” multi-tenant payment platform. All requests require a tenant API key carried in the Authorization header.",
    "contact": {
      "name": "ConfluxPay Engineering",
      "email": "engineering@conflux.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://conflux.com/license"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8089",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/admin/risk/rules/{id}": {
      "put": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "updateRule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRiskRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultRiskRuleDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "disableRule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/risk/profiles/{merchantId}": {
      "get": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "merchantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultMerchantRiskProfileDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "upsertProfile",
        "parameters": [
          {
            "name": "merchantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertMerchantRiskProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultMerchantRiskProfileDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/risk/rules": {
      "get": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "listRules",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListRiskRuleDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "createRule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRiskRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultRiskRuleDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/risk/cases/{evaluationId}/reject": {
      "post": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "rejectCase",
        "parameters": [
          {
            "name": "evaluationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/risk/cases/{evaluationId}/approve": {
      "post": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "approveCase",
        "parameters": [
          {
            "name": "evaluationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/risk/blacklist": {
      "get": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "listBlacklistEntries",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListBlacklistEntryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "addBlacklistEntry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBlacklistEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultBlacklistEntryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/risk/cases": {
      "get": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "listCases",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "FLAG",
              "enum": [
                "ALLOW",
                "FLAG",
                "BLOCK"
              ]
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListRiskCaseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/risk/blacklist/{id}": {
      "delete": {
        "tags": [
          "admin-risk-controller-impl"
        ],
        "operationId": "removeBlacklistEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UpdateRiskRuleRequest": {
        "required": [
          "action",
          "expression",
          "scoreWeight"
        ],
        "type": "object",
        "properties": {
          "expression": {
            "type": "string"
          },
          "scoreWeight": {
            "type": "integer",
            "format": "int32"
          },
          "action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "FLAG",
              "BLOCK"
            ]
          }
        }
      },
      "ApiResultMeta": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "RESOURCE_NOT_FOUND",
              "RESOURCE_ALREADY_EXISTS",
              "INVALID_OPERATION_STATE",
              "INSUFFICIENT_FUNDS",
              "IDEMPOTENCY_CONFLICT",
              "MFS_ADAPTER_FAILURE",
              "VENDOR_DOWN",
              "QUOTA_EXCEEDED",
              "RISK_REJECTED",
              "TOKEN_EXPIRED",
              "INTERNAL_ERROR"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResultRiskRuleDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RiskRuleDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RiskRuleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "expression": {
            "type": "string"
          },
          "scoreWeight": {
            "type": "integer",
            "format": "int32"
          },
          "action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "FLAG",
              "BLOCK"
            ]
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "UpsertMerchantRiskProfileRequest": {
        "required": [
          "trustLevel"
        ],
        "type": "object",
        "properties": {
          "trustLevel": {
            "type": "string",
            "enum": [
              "NEW",
              "VERIFIED",
              "TRUSTED",
              "VIP"
            ]
          },
          "customLimits": {
            "type": "string"
          }
        }
      },
      "ApiResultMerchantRiskProfileDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MerchantRiskProfileDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "MerchantRiskProfileDto": {
        "type": "object",
        "properties": {
          "merchantId": {
            "type": "string",
            "format": "uuid"
          },
          "trustLevel": {
            "type": "string",
            "enum": [
              "NEW",
              "VERIFIED",
              "TRUSTED",
              "VIP"
            ]
          },
          "customLimits": {
            "type": "string"
          }
        }
      },
...
      "ApiResultVoid": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
...
      "CreateRiskRuleRequest": {
        "required": [
          "action",
          "expression",
          "name",
          "scoreWeight"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "expression": {
            "type": "string"
          },
          "scoreWeight": {
            "type": "integer",
            "format": "int32"
          },
          "action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "FLAG",
              "BLOCK"
            ]
          }
        }
      },
      "AddBlacklistEntryRequest": {
        "required": [
          "reason",
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "EMAIL",
              "IP",
              "MERCHANT"
            ]
          },
          "value": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResultBlacklistEntryDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BlacklistEntryDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "BlacklistEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "EMAIL",
              "IP",
              "MERCHANT"
            ]
          },
          "value": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResultListRiskRuleDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RiskRuleDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "ApiResultListRiskCaseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RiskCaseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "RiskCaseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "merchantId": {
            "type": "string",
            "format": "uuid"
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          },
          "decision": {
            "type": "string",
            "enum": [
              "ALLOW",
              "FLAG",
              "BLOCK"
            ]
          },
          "triggeredRuleIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "reason": {
            "type": "string"
          },
          "evaluatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedByAdminId": {
            "type": "string",
            "format": "uuid"
          },
          "reviewDecision": {
            "type": "string"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResultListBlacklistEntryDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlacklistEntryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "description": "Tenant API key. Send as `Authorization: Bearer <api-key>`. Keys are issued per tenant in Phase 1; rotation invalidates previous keys immediately.",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  }
}